calendar: updated file for calendar access authorization
