+++ back/.eslintrc.json
@@ -22,6 +22,7 @@
     "camelcase": "off",
     "no-unused-expressions": "off",
     "no-underscore-dangle": "off",
-    "no-param-reassign": "off"
+    "no-param-reassign": "off",
+    "consistent-return": "off"
   }
 }
