+++ back/modules/calendar/authorize.js
@@ -1,0 +1,53 @@
+const fs = require('fs').promises
+const path = require('path')
+const process = require('process')
+const { authenticate } = require('@google-cloud/local-auth')
+
+// If modifying these scopes, delete token.json.
+const SCOPES = ['https://www.googleapis.com/auth/calendar']
+// The file token.json stores the user's access and refresh tokens, and is
+// created automatically when the authorization flow completes for the first
+// time.
+const TOKEN_PATH = path.join(process.cwd(), '/modules/calendar/calendar.google.token.json')
+const CREDENTIALS_PATH = path.join(
+  process.cwd(),
+  '/modules/calendar/calendar.google.credentials.json'
+)
+
+/**
+ * Serializes credentials to a file compatible with GoogleAUth.fromJSON.
+ *
+ * @param {OAuth2Client} client
+ * @return {Promise<void>}
+ */
+async function saveToken(client) {
+  const content = await fs.readFile(CREDENTIALS_PATH)
+  const keys = JSON.parse(content)
+  const key = keys.installed || keys.web
+  const payload = JSON.stringify({
+    type: 'authorized_user',
+    client_id: key.client_id,
+    client_secret: key.client_secret,
+    refresh_token: client.credentials.refresh_token,
+  })
+  await fs.writeFile(TOKEN_PATH, payload)
+}
+
+/**
+ * Request for authorization to call APIs.
+ *
+ */
+async function authorize() {
+  const client = await authenticate({
+    scopes: SCOPES,
+    keyfilePath: CREDENTIALS_PATH,
+  })
+  if (client.credentials) {
+    await saveToken(client)
+  }
+  return client
+}
+
+authorize()
+  .then(() => console.log('Token saved.'))
+  .catch((err) => console.log('Authorization failed. \n', err))
