+++ back/modules/calendar/calendar.googleAPI.js
@@ -5,17 +5,70 @@ const process = require('process')
 const { authenticate } = require('@google-cloud/local-auth')
 const { google } = require('googleapis')
 const logger = require('../../utils/logger')
+
 // If modifying these scopes, delete token.json.
 const SCOPES = ['https://www.googleapis.com/auth/calendar']
-// The file token.json stores the user's access and refresh tokens, and is
-// created automatically when the authorization flow completes for the first
-// time.
 
+// The file token.json stores the user's access and refresh tokens, and is
+// created automatically when the authorization flow completes for the first time.
 const TOKEN_PATH = path.join(process.cwd(), '/modules/calendar/calendar.google.token.json')
 const CREDENTIALS_PATH = path.join(
   process.cwd(),
   '/modules/calendar/calendar.google.credentials.json'
 )
+
+/**
+ * Reads previously authorized token from the saved file.
+ *
+ * @return {Promise<OAuth2Client|null>}
+ */
+async function loadSavedTokenIfExist() {
+  try {
+    const content = await fs.readFile(TOKEN_PATH)
+    const credentials = JSON.parse(content)
+    return google.auth.fromJSON(credentials)
+  } catch (err) {
+    return null
+  }
+}
+
+/**
+ * Serializes token to a file compatible with GoogleAUth.fromJSON.
+ *
+ * @param {OAuth2Client} client
+ * @return {Promise<void>}
+ */
+async function saveToken(client) {
+  const content = await fs.readFile(CREDENTIALS_PATH)
+  const keys = JSON.parse(content)
+  const key = keys.installed || keys.web
+  const payload = JSON.stringify({
+    type: 'authorized_user',
+    client_id: key.client_id,
+    client_secret: key.client_secret,
+    refresh_token: client.credentials.refresh_token,
+  })
+  await fs.writeFile(TOKEN_PATH, payload)
+}
+
+/**
+ * Load or request or authorization to call APIs.
+ *
+ */
+async function authorize() {
+  let client = await loadSavedTokenIfExist()
+  if (client) {
+    return client
+  }
+  client = await authenticate({
+    scopes: SCOPES,
+    keyfilePath: CREDENTIALS_PATH,
+  })
+  if (client.credentials) {
+    await saveToken(client)
+  }
+  return client
+}
 
 /**
  * Adds event to the Google calendar. Before that checks OAuth credentials. Mostly boilerplate from Google Docs example.
@@ -82,59 +135,6 @@ function deleteEventFromCalendar(eventId) {
 
     authorize().then(deleteEvent)
   })
-}
-
-/**
- * Load or request or authorization to call APIs.
- *
- */
-async function authorize() {
-  let client = await loadSavedCredentialsIfExist()
-  if (client) {
-    return client
-  }
-  client = await authenticate({
-    scopes: SCOPES,
-    keyfilePath: CREDENTIALS_PATH,
-  })
-  if (client.credentials) {
-    await saveCredentials(client)
-  }
-  return client
-}
-
-/**
- * Serializes credentials to a file compatible with GoogleAUth.fromJSON.
- *
- * @param {OAuth2Client} client
- * @return {Promise<void>}
- */
-async function saveCredentials(client) {
-  const content = await fs.readFile(CREDENTIALS_PATH)
-  const keys = JSON.parse(content)
-  const key = keys.installed || keys.web
-  const payload = JSON.stringify({
-    type: 'authorized_user',
-    client_id: key.client_id,
-    client_secret: key.client_secret,
-    refresh_token: client.credentials.refresh_token,
-  })
-  await fs.writeFile(TOKEN_PATH, payload)
-}
-
-/**
- * Reads previously authorized credentials from the save file.
- *
- * @return {Promise<OAuth2Client|null>}
- */
-async function loadSavedCredentialsIfExist() {
-  try {
-    const content = await fs.readFile(TOKEN_PATH)
-    const credentials = JSON.parse(content)
-    return google.auth.fromJSON(credentials)
-  } catch (err) {
-    return null
-  }
 }
 
 /**
